Move alert type field and selector components to child directory
